requirements.txt
pandas 
numpy
scikit-learn
pyyaml
python-dotenv
gunicorn
boto3
flask
